Shubhra/Organize RoomIO / Transcription Logic (#443)
