+++ agents/src/voice/agent_session.ts
@@ -16,8 +16,8 @@ import type { Agent } from './agent.js';
 import { AgentActivity } from './agent_activity.js';
 import type { _TurnDetector } from './audio_recognition.js';
 import type { UserState } from './events.js';
-import { type ParticipantAudioOutput, RoomIO } from './room_io.js';
-import type { TextOutput } from './room_io/_output.js';
+import type { ParticipantAudioOutput, TextOutput } from './room_io/index.js';
+import { RoomIO } from './room_io/index.js';
 
 export type AgentState = 'initializing' | 'thinking' | 'listening' | 'speaking';
 export interface VoiceOptions {
