+++ agents/src/voice/room_io/room_io.ts
@@ -1,0 +1,178 @@
+// SPDX-FileCopyrightText: 2025 LiveKit, Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+import { type AudioFrame, type Participant, type Room } from '@livekit/rtc-node';
+import {
+  AudioStream,
+  type RemoteTrack,
+  RoomEvent,
+  TrackKind,
+  TrackPublishOptions,
+  TrackSource,
+} from '@livekit/rtc-node';
+import type { ReadableStream } from 'node:stream/web';
+import { ATTRIBUTE_PUBLISH_ON_BEHALF } from '../../constants.js';
+import { log } from '../../log.js';
+import { DeferredReadableStream } from '../../stream/deferred_stream.js';
+import { Future } from '../../utils.js';
+import {
+  type AgentSession,
+  AgentSessionEvent,
+  type UserInputTranscribedEvent,
+} from '../agent_session.js';
+import {
+  ParalellTextOutput,
+  ParticipantAudioOutput,
+  ParticipantLegacyTranscriptionOutput,
+  ParticipantTranscriptionOutput,
+} from './_output.js';
+
+export class RoomIO {
+  private agentSession: AgentSession;
+  private participantAudioInputStream: ReadableStream<AudioFrame>;
+
+  private room: Room;
+
+  private _deferredAudioInputStream = new DeferredReadableStream<AudioFrame>();
+  private participantAudioOutput?: ParticipantAudioOutput;
+  private userTranscriptOutput?: ParalellTextOutput;
+  private agentTranscriptOutput?: ParalellTextOutput;
+
+  private participantIdentity?: string;
+  private participantAvailableFuture: Future<Participant> = new Future();
+  private roomConnectedFuture: Future<void> = new Future();
+
+  private logger = log();
+
+  constructor(
+    agentSession: AgentSession,
+    room: Room,
+    private readonly sampleRate: number,
+    private readonly numChannels: number,
+  ) {
+    this.agentSession = agentSession;
+    this.room = room;
+    this.participantAudioInputStream = this._deferredAudioInputStream.stream;
+  }
+
+  private onTrackSubscribed = (track: RemoteTrack) => {
+    if (track.kind === TrackKind.KIND_AUDIO) {
+      this._deferredAudioInputStream.setSource(
+        new AudioStream(track, {
+          // TODO(AJS-41) remove hardcoded sample rate
+          sampleRate: 16000,
+          numChannels: 1,
+        }),
+      );
+    }
+  };
+
+  private async initTask() {
+    await this.roomConnectedFuture.await;
+
+    for (const participant of this.room.remoteParticipants.values()) {
+      this.onParticipantConnected(participant);
+    }
+
+    const participant = await this.participantAvailableFuture.await;
+
+    // init user outputs
+    this.updateTranscriptionOutput(this.userTranscriptOutput, participant.identity);
+
+    // init agent outputs
+    this.updateTranscriptionOutput(
+      this.agentTranscriptOutput,
+      this.room.localParticipant?.identity,
+    );
+
+    await this.participantAudioOutput?.start();
+  }
+
+  private onParticipantConnected(participant: Participant) {
+    if (this.participantAvailableFuture.done) {
+      return;
+    }
+
+    if (this.participantIdentity) {
+      if (participant.identity !== this.participantIdentity) {
+        return;
+      }
+    } else if (
+      // otherwise, skip participants that are marked as publishing for this agent
+      participant.attributes?.[ATTRIBUTE_PUBLISH_ON_BEHALF] === this.room.localParticipant?.identity
+    ) {
+      return;
+    }
+
+    // TODO(AJS-105): allow user to specify accepted participany kinds
+
+    this.participantAvailableFuture.resolve(participant);
+  }
+
+  private onUserInputTranscribed = (ev: UserInputTranscribedEvent) => {
+    this.logger.debug({ ev }, 'user input transcribed');
+    this.userTranscriptOutput?.captureText(ev.transcript);
+    if (ev.isFinal) {
+      this.userTranscriptOutput?.flush();
+    }
+  };
+
+  private createTranscriptionOutput(options: { isDeltaStream: boolean; participant?: string }) {
+    return new ParalellTextOutput([
+      new ParticipantLegacyTranscriptionOutput(
+        this.room,
+        options.isDeltaStream,
+        options.participant,
+      ),
+      new ParticipantTranscriptionOutput(this.room, options.isDeltaStream, options.participant),
+    ]);
+  }
+
+  private updateTranscriptionOutput(output?: ParalellTextOutput, participant?: string) {
+    if (!output) {
+      return;
+    }
+
+    for (const sink of output._sinks) {
+      if (
+        sink instanceof ParticipantLegacyTranscriptionOutput ||
+        sink instanceof ParticipantTranscriptionOutput
+      ) {
+        sink.setParticipant(participant);
+      }
+    }
+  }
+
+  start() {
+    // -- create outputs --
+    this.participantAudioOutput = new ParticipantAudioOutput(this.room, {
+      sampleRate: this.sampleRate,
+      numChannels: this.numChannels,
+      trackPublishOptions: new TrackPublishOptions({ source: TrackSource.SOURCE_MICROPHONE }),
+    });
+
+    this.userTranscriptOutput = this.createTranscriptionOutput({
+      isDeltaStream: false,
+      participant: this.participantIdentity,
+    });
+    this.agentTranscriptOutput = this.createTranscriptionOutput({
+      isDeltaStream: true,
+    });
+
+    // -- set the room event handlers --
+    this.room.on(RoomEvent.ParticipantConnected, this.onParticipantConnected);
+    this.room.on(RoomEvent.TrackSubscribed, this.onTrackSubscribed);
+    if (this.room.isConnected) {
+      this.roomConnectedFuture.resolve();
+    }
+
+    this.initTask();
+
+    // -- attatch the agent to the session --
+    this.agentSession.audioInput = this.participantAudioInputStream;
+    this.agentSession.audioOutput = this.participantAudioOutput;
+    this.agentSession._transcriptionOutput = this.agentTranscriptOutput;
+
+    this.agentSession.on(AgentSessionEvent.UserInputTranscribed, this.onUserInputTranscribed);
+  }
+}
