+++ agents/src/voice/transcription/index.ts
@@ -1,5 +1,4 @@
 // SPDX-FileCopyrightText: 2025 LiveKit, Inc.
 //
 // SPDX-License-Identifier: Apache-2.0
-export * from './_output.js';
-export * from './room_io.js';
+export * from './_utils.js';
